Move create-review to advertisement type
