+++ cms/api/order/config/routes.json
@@ -39,14 +39,6 @@
       "config": {
         "policies": []
       }
-    },
-    {
-      "method": "POST",
-      "path": "/create-review/:key",
-      "handler": "order.createReview",
-      "config": {
-        "policies": []
-      }
     }
   ]
 }
